+++ examples/scripts/run_wgs_single.sh
@@ -14,6 +14,7 @@ genome: hg38
 resource: cbicall-germline-resources-v1
 input_dir: "$INPUT_DIR"
 cleanup_bam: false
+export_mtdna_bam: true
 EOF
 
 cbicall run -p wgs_single.yaml -t 4